create config local, modify template bunddle, remove folder /core
