--- packages/core/tsconfig.json
@@ -1,28 +1,0 @@
-{
-  "compilerOptions": {
-    /* Basic Options */
-    "target": "es5",
-    "module": "commonjs",
-    "lib": ["ESNext"],
-    "declaration": true,
-    "removeComments": true,
-    "composite": true,
-    "outDir": "dist",
-    "rootDir": "src",
-
-    /* Strict Type-Checking Options */
-    "strict": true,
-
-    /* Module Resolution Options */
-    "moduleResolution": "node",
-    "esModuleInterop": true,
-    // "baseUrl": "",
-    // "paths": {},
-
-    /* Advanced Options */
-    "forceConsistentCasingInFileNames": true,
-    "skipLibCheck": true
-  },
-  "include": ["src/**/*.ts"],
-  "exclude": ["node_modules"]
-}
