+++ packages/css-battle-api/src/@types/main.d.ts
@@ -1,0 +1,10 @@
+type TConfig = {
+  urlBase: string
+  urlProxy: string
+}
+
+type TProxy = import('../types').TConfig['proxy']
+
+/************************************ DECLARATIONS **************************************/
+
+declare var config: TConfig
