+++ packages/css-battle-api/src/helpers/loadConfig.ts
@@ -1,0 +1,8 @@
+export function loadConfig() {
+  const config: TConfig = {
+    urlBase: 'https://cssbattle.dev',
+    urlProxy: 'https://css-battle-proxy.herokuapp.com/'
+  }
+
+  global.config = config
+}
